pyproject.toml
[project]
name = "langgraph-basic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asttokens==3.0.0",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "colorama==0.4.6",
    "comm==0.2.2",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "executing==2.2.0",
    "greenlet==3.2.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "ipykernel==6.29.5",
    "ipython==9.2.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "langchain[openai]==0.3.25",
    "langchain-core==0.3.61",
    "langchain-ollama==0.3.3",
    "langchain-tavily>=0.1.6",
    "langchain-text-splitters==0.3.8",
    "langgraph==0.4.5",
    "langgraph-checkpoint==2.0.26",
    "langgraph-prebuilt==0.2.0",
    "langgraph-sdk==0.1.70",
    "langsmith==0.3.42",
    "matplotlib-inline==0.1.7",
    "nest-asyncio==1.6.0",
    "ollama==0.4.8",
    "orjson==3.10.18",
    "ormsgpack==1.9.1",
    "packaging==24.2",
    "parso==0.8.4",
    "platformdirs==4.3.8",
    "prompt-toolkit==3.0.51",
    "psutil==7.0.0",
    "pure-eval==0.2.3",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pygments==2.19.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pywin32==310",
    "pyyaml==6.0.2",
    "pyzmq==26.4.0",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "stack-data==0.6.3",
    "tenacity==9.1.2",
    "tornado==6.5.1",
    "traitlets==5.14.3",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.1",
    "urllib3==2.4.0",
    "wcwidth==0.2.13",
    "xxhash==3.5.0",
    "zstandard==0.23.0",
]
